+++ src/pages/account/Register.vue
